refactor: extract login/auth flow into login_flow.py
Extract all login, Auth0 form-fill, captcha, SMS/email verification, GDPR,
login-state detection, and diagnostics-capture logic from __init__.py into
a new login_flow module with module-level async functions and explicit
dependencies.

Create login_flow.py (~634 lines) with module-level async functions.
Thin delegators in __init__.py bind instance attributes to module calls.
Remove 12 dead delegator methods called only from tests; update tests
to call module functions directly.
Move LoginDetectionResult/LoginDetectionReason to login_flow.
Move auth tests to tests/unit/test_login_flow.py, update patch targets.
Move login translations to login_flow.py section in translations.de.yaml.
Fix diagnostics guard to reset per login attempt and use web attribute.
Remove is_logged_in_web; pass username explicitly.
Make cross-module functions public: login_flow internal helpers and
WebScrapingMixin helpers (dismiss_consent_banner, effective_timeout,
extract_visible_text).
